add rules list drawer
