+++ app/src/features/rules/screens/rulesList/components/RulesList/components/RuleSelectionListDrawer/RuleSelectionListDrawer.tsx
@@ -1,0 +1,43 @@
+import React from "react";
+import { Drawer, DrawerProps } from "antd";
+import { RuleSelectionList } from "../RuleSelectionList/RuleSelectionList";
+import "./RuleSelectionListDrawer.scss";
+
+interface Props {
+  open: boolean;
+  onClose: () => void;
+  groupId?: string;
+  children: React.ReactNode;
+  analyticEventSource: string;
+  placement?: DrawerProps["placement"];
+}
+
+export const RuleSelectionListDrawer: React.FC<Props> = ({
+  open,
+  onClose,
+  groupId,
+  children,
+  placement = "right",
+  analyticEventSource,
+}) => {
+  return (
+    <>
+      <Drawer
+        title="Create new rule"
+        open={open}
+        placement={placement}
+        onClose={onClose}
+        className="rule-selection-list-drawer"
+      >
+        <RuleSelectionList
+          groupId={groupId}
+          source={analyticEventSource}
+          premiumPopoverPlacement="topLeft"
+          premiumIconSource="rule_dropdown"
+        />
+      </Drawer>
+
+      {children}
+    </>
+  );
+};
