+++ app/src/features/rules/screens/rulesList/components/RulesList/components/index.ts
@@ -2,6 +2,6 @@ export { GettingStarted } from "./GettingStarted";
 export { CreateTeamRuleCTA } from "./GettingStarted/CreateTeamRuleCTA";
 export { ImportFromCharlesModal, ImportFromCharlesWrapperView } from "./CharlesImporter";
 export { ImportRulesModal } from "../../../../../modals/ImportRulesModal";
-export { RuleTypesDropdownWrapper } from "./RuleTypesDropdownWrapper/RuleTypesDropdownWrapper";
+export { RuleSelectionListDrawer } from "./RuleSelectionListDrawer/RuleSelectionListDrawer";
 export { NewRuleButtonWithDropdown } from "./NewRuleButtonWithDropdown/NewRuleButtonWithDropdown";
 export { RuleSelectionList } from "./RuleSelectionList/RuleSelectionList";
