+++ app/src/features/rules/screens/rulesList/components/RulesList/rulesList.scss
@@ -17,7 +17,7 @@
 .rules-dropdown-items-container {
   width: 360px;
   height: 100%;
-  max-height: calc(100vh - 220px);
+  max-height: calc(100vh - 240px);
   display: flex;
   flex-direction: column;
   align-items: center;
